MOVEMENT SYSTEM - Change to float VelocityComponent

+++ core/src/main/java/com/fatpiggies/game/model/ecs/components/VelocityComponent.java
@@ -3,8 +3,8 @@ package com.fatpiggies.game.model.ecs.components;
 import com.badlogic.ashley.core.Component;
 
 public class VelocityComponent implements Component {
-    public double baseMaxVelocity;
-    public double currentMaxVelocity;
-    public double vx;
-    public double vy;
+    public float baseMaxVelocity;
+    public float currentMaxVelocity;
+    public float vx;
+    public float vy;
 }
